+++ HISTORY.rst
@@ -43,6 +43,8 @@
 * Move the navigation and authentication templates to their
   respective apps.
 * Add dashboard app.
+* Remove queryset slicing hack from the Document list view.
+  And slice the Recently Added Document queryset itself.
 
 3.1.11 (2019-04-XX)
 ===================